break out UI component (footer)
